Add files via upload
The only way to comfortably track your account balance with a Sparkasse account is by activating the Sparkasse Kontowecker (which is a mail service that sends you an e-mail with your current balance each day).

This script extracts your current account balance out of the e-mail (using regex) and pastes it in a Google Spreadsheet with the current date. By automating this process by running the Google Apps script on a daily basis, you'll get an uptodate overview of your bank account without having to check manually your account.